requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
lxml==4.9.1
magic-filter==1.0.9
multidict==6.0.2
peewee==3.15.4
Pillow==9.3.0
psycopg2-binary==2.9.5
python-decouple==3.6
python-docx==0.8.11
pytz==2022.6
requests==2.28.1
urllib3==1.26.13
yarl==1.8.1
